Rename CLI modules and add placeholders
